make the line field in sl_call_frame a size_t*

+++ inc/slash/vm.h
@@ -188,7 +188,7 @@ typedef struct sl_vm_frame {
         } c_call_frame;
         struct {
             sl_vm_section_t* section;
-            int* line;
+            size_t* line;
         } sl_call_frame;
         struct {
             jmp_buf env;
